# 25x speed improvement in tokenizing jquery.min: maxCachedIndex should not dicrease due to a cache miss