@zcbenz 2015-09-21
This change looks good to me, and the improvement is awesome!

> I am not familiar enough with v8, but I had to do the following unrelated change to even get the string cache to work (IsSame was always false for me in this node version). I ran both benchmarks with this change in, without this I gave up after 20min of waiting:

Current implementation of `OnigStringContext::IsSame` is wrong, can you replace it with your pasted one? We don't need to compare the length of string though, the `StrictEquals` method already included many optimizations.

> I would be very thankful if you cherry pick this to master and also publish a new v4 on npm with this change, if you agree with my reasoning.

It sounds good to me.